backend/requirements.txt
# Django & Core
Django==5.1.5
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-oauth-toolkit==3.0.1

# Database
psycopg2-binary==2.9.10

# Admin Theme
django-unfold==0.44.0

# Email
django-templated-mail==1.1.1

# Security
django-ratelimit==4.1.0

# Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
factory-boy==3.3.1

# Server
gunicorn==23.0.0
